nebula_decoders. cleanup decoders

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/pandar_qt_128.hpp
@@ -15,8 +15,8 @@ namespace pandar_qt_128
 {
 constexpr uint16_t MAX_AZIMUTH_STEPS = 900;  // High Res mode
 //constexpr float DISTANCE_UNIT = 0.004f;       // 4mm
-constexpr float MIN_RANGE = 0.05;
-constexpr float MAX_RANGE = 50.0;
+constexpr double MIN_RANGE = 0.05;
+constexpr double MAX_RANGE = 50.0;
 
 // Head
 constexpr size_t HEAD_SIZE = 12;
